// src/lib/images.generated.ts
/* Generated by scripts/build-images.mjs — do not edit. */
export const IMAGES = {
  "waiting-room": {
    "widths": [
      560,
      1120
    ],
    "width": 560,
    "height": 373,
    "lqip": "data:image/webp;base64,UklGRoIAAABXRUJQVlA4IHYAAABQBACdASoYABAAPxFysFAsJqSisAgBgCIJZQC/OCPLVupsss7E8LXV6QCAAPuGZKivrukTFGkso+6LXs+2eOHvCZkTNLgG4IGcSPiT96Vtk/yzZ4v/D+UnApt+KMEYl/sepuxfPj1mlYPtaHu/uCprmz6jeAAA"
  },
  "arcade": {
    "widths": [
      480,
      960
    ],
    "width": 480,
    "height": 720,
    "lqip": "data:image/webp;base64,UklGRrgAAABXRUJQVlA4IKwAAABwBQCdASoYACQAPxFws1CsJiSiqrgN+YAiCWMAvKQisFIdgARMNW+fjathVw6zlojMBdzAAP71zU0aXEox1DZ4YznBYWcaqrY8LgkI4k2m7MpQdVp7G3E2xNcJUQqmw3tjMIPrYqhCF1PykH0GTkkybQbXgygxPGKDfdcp0bl99XEbsx/GFtVHQkL25231fVbn65vsUFm8N4STDoGh3btD2NHaSTg+UofoBRgA"
  },
  "market": {
    "widths": [
      720,
      1440
    ],
    "width": 720,
    "height": 480,
    "lqip": "data:image/webp;base64,UklGRnIAAABXRUJQVlA4IGYAAADQAwCdASoYABAAPxFysFAsJqSisAgBgCIJZQC+SCDpLLgXY8Nk+YAA/uvjWQfuRcxLGyi2nL0ZosLy/lEQIZp8ZTxNYnOX9rmgYppraSnd3BJXI6gIRayOBhmeF0IjucBsC/ygAAA="
  },
  "kopitiam": {
    "widths": [
      560,
      1120
    ],
    "width": 560,
    "height": 560,
    "lqip": "data:image/webp;base64,UklGRpQAAABXRUJQVlA4IIgAAACwBACdASoYABgAPxF2s1KsJySisBgIAYAiCWcAACnTU7ExpVHWT6Dynsy/PQgAAParZsrujYD1ooBLw+nKVDhGMTCsRyBWqJXCqu+/OsHONPfoKJ6RsHw2GOGNn8oGAjZqHWwEhumUMCliqwL+9f74lDBgkmojLeVcZXkkOhOU8xhS5DOsAAAA"
  },
  "bus-stop": {
    "widths": [
      720,
      1440
    ],
    "width": 720,
    "height": 480,
    "lqip": "data:image/webp;base64,UklGRngAAABXRUJQVlA4IGwAAACQAwCdASoYABAAPxFysVAsJqSisAgBgCIJZwC/OAysETqSQ6n8AP4UyvlZFwFAHqLG+kUcJfVSMv0k2TNbGCEUzaNCgOfPc7m/0N+ThmOsDf56bVeTaE+AUoaEEz5Irtjmx4GGUpCu+UK8OAA="
  },
  "hands": {
    "widths": [
      560,
      1120
    ],
    "width": 560,
    "height": 560,
    "lqip": "data:image/webp;base64,UklGRowAAABXRUJQVlA4IIAAAAAwBQCdASoYABgAPxF+tVWsJ6SjKAgBgCIJZQAAIsS0iwd1ZPI18wAhVM/3Ev59UFjGAAD8ovHvagKjPG7AA9cDLN07QfZxS5Dxj8oLuxj2q1sKcslB2F//J/3NY65HlJWcnlrrsL1QvC7shl1E78sgs19vkyyxiN/nuFGg2qgAAA=="
  },
  "corridor": {
    "widths": [
      480,
      960
    ],
    "width": 480,
    "height": 720,
    "lqip": "data:image/webp;base64,UklGRroAAABXRUJQVlA4IK4AAAAQBgCdASoYACQAPxF8s1SsJ6QjJWzJgCIJYwC2FDaVFfSxB56UYfpX32z1XDyrY2s+KZLDI4UXdAAA/uwMLsc0RdwsTAdF+OxNaYkKQmPkICs61co8XL9f3V0a/hN4RkrgitXsAwhTsxEE9NM8KGNNRRrU3dk8gIO0M4UWrkcKo07z5Dn1QxpSyJw899m47j5iqEFnOYwBU531XdPpOKoP4avMW8xaV8pYivUKwAA="
  },
  "chair-ticket": {
    "widths": [
      720,
      1440
    ],
    "width": 720,
    "height": 480,
    "lqip": "data:image/webp;base64,UklGRnwAAABXRUJQVlA4IHAAAAAQBACdASoYABAAPxF2sVCsJySisAgBgCIJQBibApiBi7q+129tG+JtgAD+ufmFffUHQMIXCUj2VeXOY9PyX3J0K9vJiTRWsBi8FE1O9z064y/9nS+0+osnwKp+FqPOYZA7k7EwehWznAfv6CI9nAAA"
  }
} as const;